refactor: 인터페이스와 구현체로 분리하기
